changed float64 types, additional double_t to float32, floating

+++ hdbscan/robust_single_linkage_.py
@@ -49,7 +49,7 @@ def _rsl_prims_kdtree(X, k=5, alpha=1.4142135623730951, metric='euclidean',
 
     # The Cython routines used require contiguous arrays
     if not X.flags['C_CONTIGUOUS']:
-        X = np.array(X, dtype=np.double, order='C')
+        X = np.array(X, dtype=X.dtype, order='C')
 
     dim = X.shape[0]
     k = min(dim - 1, k)
@@ -73,7 +73,7 @@ def _rsl_prims_balltree(X, k=5, alpha=1.4142135623730951, metric='euclidean',
 
     # The Cython routines used require contiguous arrays
     if not X.flags['C_CONTIGUOUS']:
-        X = np.array(X, dtype=np.double, order='C')
+        X = np.array(X, dtype=X.dtype, order='C')
 
     dim = X.shape[0]
     k = min(dim - 1, k)
